Move init for api client to init phase - extract requesthandler from main file
