+++ src/app/uitkeringen/tests/uitkering.test.ts
@@ -2,7 +2,7 @@ import { DynamoDBClient, GetItemCommandOutput } from '@aws-sdk/client-dynamodb';
 import { SecretsManagerClient, GetSecretValueCommandOutput } from '@aws-sdk/client-secrets-manager';
 import { mockClient } from 'jest-aws-client-mock';
 import { FileApiClient } from '../FileApiClient';
-import * as lambda from '../index';
+import { requestHandler } from '../requestHandler';
 
 beforeAll(() => {
   global.console.log = jest.fn();
@@ -46,7 +46,7 @@ test('Returns 200', async () => {
   secretsMock.mockImplementation(() => output);
   const client = new FileApiClient();
   const dynamoDBClient = new DynamoDBClient({});
-  const result = await lambda.requestHandler('session=12345', client, dynamoDBClient);
+  const result = await requestHandler('session=12345', client, dynamoDBClient);
   expect(result.statusCode).toBe(200);
 });
 
@@ -57,8 +57,8 @@ test('Shows overview page', async () => {
   };
   secretsMock.mockImplementation(() => output);
   const client = new FileApiClient();
-  const dynamoDBClient = new DynamoDBClient({});
-  const result = await lambda.requestHandler('session=12345', client, dynamoDBClient);
+  const dynamoDBClient = new DynamoDBClient({ region: 'eu-west-1' });
+  const result = await requestHandler('session=12345', client, dynamoDBClient);
   expect(result.body).toMatch('Mijn Uitkering');
   expect(result.body).toMatch('Participatiewet');
 });
